refactor: move sql adapter files into adapter folder for postgres.
