+++ packages/serverpod/lib/src/database/adapters/postgres/database_connection.dart
@@ -5,16 +5,16 @@ import 'package:retry/retry.dart';
 import 'package:serverpod/src/database/concepts/columns.dart';
 import 'package:serverpod/src/database/concepts/table_relation.dart';
 import 'package:serverpod/src/database/sql_query_builder.dart';
-import 'package:serverpod/src/database/database_result.dart';
+import 'package:serverpod/src/database/adapters/postgres/database_result.dart';
 import 'package:serverpod/src/database/concepts/includes.dart';
 import 'package:serverpod/src/database/concepts/order.dart';
 import 'package:serverpod/src/database/concepts/transaction.dart';
 
-import '../generated/protocol.dart';
-import '../server/session.dart';
-import 'database_pool_manager.dart';
-import 'concepts/expressions.dart';
-import 'concepts/table.dart';
+import '../../../generated/protocol.dart';
+import '../../../server/session.dart';
+import '../../database_pool_manager.dart';
+import '../../concepts/expressions.dart';
+import '../../concepts/table.dart';
 
 /// A connection to the database. In most cases the [Database] db object in
 /// the [Session] object should be used when connecting with the database.
